wjdqhrnwhProject/backend/requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub
chromadb
yfinance
pandas
numpy
fastapi
uvicorn
python-dotenv
openai
tiktoken